Fix tax rate storage and broken indices, and auto-calculate invoice lines.
- Modified `create_taxes_table` migration to use `decimal(10, 4)` for `rate` column, ensuring fractional tax rates are stored correctly (e.g., 0.1500 for 15%).
- Fixed `add_inventory_performance_indexes` migration to remove indices targeting non-existent columns on `stock_moves` table, resolving schema errors.
- Implemented `InvoiceLineObserver::saving` to automatically calculate `subtotal` and `total_line_tax` using `Brick\Money` for precise financial calculations.
- Cleaned up database with `migrate:fresh` to apply fixes properly in development environment.

+++ database/migrations/2025_07_21_144434_create_taxes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('company_id')->constrained('companies');
             $table->foreignId('tax_account_id')->constrained('accounts');
             $table->json('name');
-            $table->unsignedBigInteger('rate'); // e.g., 0.15000 for 15%
+            $table->decimal('rate', 10, 4); // e.g., 0.1500 for 15%
             $table->string('type'); // 'sales', 'purchase', 'both'
             $table->boolean('is_active')->default(true);
             $table->boolean('is_recoverable')->default(true); // Whether tax can be deducted as input tax or should be capitalized
